server/requirements.txt
# server/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
aiosqlite==0.20.0
python-dotenv==1.0.1
google-generativeai==0.7.0
easyocr==1.7.1
python-multipart==0.0.9